+++ src/sourcemaps/sourcemaps-wizard.ts
@@ -16,15 +16,20 @@ import { SourceMapUploadToolConfigurationOptions } from './tools/types';
 import { configureVitePlugin } from './tools/vite';
 import { configureSentryCLI } from './tools/sentry-cli';
 import { configureWebPackPlugin } from './tools/webpack';
+import { configureTscSourcemapGenerationFlow } from './tools/tsc';
 import { configureRollupPlugin } from './tools/rollup';
 
 interface SourceMapsWizardOptions {
   promoCode?: string;
 }
 
-type SupportedBundlers = 'webpack' | 'vite' | 'rollup' | 'esbuild';
-type SupportedTools = 'sentry-cli';
-type SupportedBundlersTools = SupportedBundlers | SupportedTools;
+type SupportedTools =
+  | 'webpack'
+  | 'vite'
+  | 'rollup'
+  | 'esbuild'
+  | 'tsc'
+  | 'sentry-cli';
 
 export async function runSourcemapsWizard(
   options: SourceMapsWizardOptions,
@@ -125,8 +130,8 @@ SENTRY_AUTH_TOKEN=${apiKeys.token}
 `);
 }
 
-async function askForUsedBundlerTool(): Promise<SupportedBundlersTools> {
-  const selectedTool: SupportedBundlersTools | symbol = await abortIfCancelled(
+async function askForUsedBundlerTool(): Promise<SupportedTools> {
+  const selectedTool: SupportedTools | symbol = await abortIfCancelled(
     clack.select({
       message: 'Which bundler or build tool are you using?',
       options: [
@@ -139,6 +144,11 @@ async function askForUsedBundlerTool(): Promise<SupportedBundlersTools> {
           label: 'Vite',
           value: 'vite',
           hint: 'Configure source maps upload using Vite',
+        },
+        {
+          label: 'tsc',
+          value: 'tsc',
+          hint: 'Configure source maps when using tsc as build tool',
         },
         {
           label: 'Rollup',
@@ -164,7 +174,7 @@ async function askForUsedBundlerTool(): Promise<SupportedBundlersTools> {
 }
 
 async function startToolSetupFlow(
-  selctedTool: SupportedBundlersTools,
+  selctedTool: SupportedTools,
   options: SourceMapUploadToolConfigurationOptions,
 ): Promise<void> {
   switch (selctedTool) {
@@ -174,10 +184,12 @@ async function startToolSetupFlow(
     case 'webpack':
       await configureWebPackPlugin(options);
       break;
+    case 'tsc':
+      await configureSentryCLI(options, configureTscSourcemapGenerationFlow);
+      break;
     case 'rollup':
       await configureRollupPlugin(options);
       break;
-    // TODO: implement other bundlers
     default:
       await configureSentryCLI(options);
       break;