+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- feat(sourcemaps): Add Rollup plugin flow #325
+- feat(sourcemaps): Add `tsc` flow (#324)
+- feat(sourcemaps): Add Rollup plugin flow (#325)
 
 ## 3.3.2